+++ src/renderer/components/dialogs/Settings.tsx
@@ -22,6 +22,8 @@ import SettingsStoreInstance, {
   SettingsStoreState,
   useSettingsStore,
 } from '../../stores/settings'
+import { runtime } from '../../runtime'
+import { donationUrl } from '../../../shared/constants'
 
 export function flipDeltaBoolean(value: string) {
   return value === '1' ? '0' : '1'
@@ -37,7 +39,7 @@ export function SettingsButton(props: any) {
 }
 
 export function SettingsIconButton(props: any) {
-  const { onClick, iconName, children, ...otherProps } = props
+  const { onClick, iconName, children, isLink, ...otherProps } = props
   return (
     <div className='SettingsIconButton' onClick={onClick}>
       <div
@@ -48,6 +50,14 @@ export function SettingsIconButton(props: any) {
         }}
       ></div>
       <button {...otherProps}>{children}</button>
+      {isLink && (
+        <div
+          className='Icon'
+          style={{
+            WebkitMask: 'url(../images/icons/open_in_new.svg) no-repeat center',
+          }}
+        ></div>
+      )}
     </div>
   )
 }
@@ -236,6 +246,13 @@ export default function Settings(props: DialogProps) {
                   onClick={() => setSettingsMode('experimental_features')}
                 >
                   {tx('pref_experimental_features')}
+                </SettingsIconButton>
+                <SettingsIconButton
+                  iconName='favorite'
+                  onClick={() => runtime.openLink(donationUrl)}
+                  isLink
+                >
+                  {tx('pref_donate')}
                 </SettingsIconButton>
               </Card>
             </DeltaDialogBody>