+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Added
 - clear webxdc browser data on webxdc instance deletion
+- add donate link in settings
 
 ## Changed
 - start migrating to jsonrpc api
 - Update translations (22.09.2022)
 - click on selected chat in chatlist now goes to bottom or first unread message
+- remember last path in "save as" dialog
 
 ## Fixed
 - allow scanning of certain qr code types on welcome screen (account, url and text)